requirements.txt
Flask==2.0.2
google-cloud-datastore==2.3.0
google-auth-oauthlib==0.4.6
google-auth==2.3.3
requests-oauthlib==1.3.0
gunicorn==20.1.0